// apps/media-server/src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const OutputProtocol = {
  UDP: 'UDP',
  SRT: 'SRT',
  RTMP: 'RTMP'
} as const

export type OutputProtocol = (typeof OutputProtocol)[keyof typeof OutputProtocol]


export const BroadcastSessionState = {
  STARTING: 'STARTING',
  RUNNING: 'RUNNING',
  STOPPING: 'STOPPING',
  COMPLETED: 'COMPLETED',
  FAILED: 'FAILED',
  STOPPED: 'STOPPED'
} as const

export type BroadcastSessionState = (typeof BroadcastSessionState)[keyof typeof BroadcastSessionState]